requirements_scbm_relaxed2.txt
annotated-types
antlr4-python3-runtime
appnope
asttokens
certifi
charset-normalizer
click
comm
contourpy
cycler
debugpy
decorator
executing
fonttools
gitdb
GitPython
hf-xet
huggingface_hub
hydra-core
idna
ipykernel
ipython
ipython_pygments_lexers
jedi
joblib
jupyter_client
jupyter_core
kiwisolver
matplotlib
matplotlib-inline
narwhals
nest-asyncio
omegaconf
pandas
parso
pexpect
platformdirs
plotly
prompt_toolkit
protobuf
psutil
ptyprocess
pure_eval
pydantic
pydantic_core
Pygments
pyparsing
python-dateutil
pytorch-minimize
PyYAML
pyzmq
regex
requests
safetensors
scikit-learn
scipy
seaborn
sentry-sdk
six
smmap
stack-data
threadpoolctl
tokenizers
torchmetrics
tornado
tqdm
traitlets
transformers
typing-inspection
urllib3
wandb
wcwidth
